# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordPayload": {
        "properties": {
          "current_password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordPayload",
        "type": "object"
      },
...
      "CrawlIndexPayload": {
        "properties": {
          "crawl_all": {
            "default": false,
            "title": "Crawl All",
            "type": "boolean"
          },
          "end_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Page"
          },
          "max_consecutive_empty": {
            "default": 2,
            "title": "Max Consecutive Empty",
            "type": "integer"
          },
          "page_delay": {
            "default": 1.0,
            "title": "Page Delay",
            "type": "number"
          },
          "start_page": {
            "default": 1,
            "title": "Start Page",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "use_cf_bypass": {
            "default": true,
            "title": "Use Cf Bypass",
            "type": "boolean"
          },
          "use_cookie": {
            "default": false,
            "title": "Use Cookie",
            "type": "boolean"
          },
          "use_proxy": {
            "default": true,
            "title": "Use Proxy",
            "type": "boolean"
          }
        },
        "required": [
          "url"
        ],
        "title": "CrawlIndexPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HtmlPayload": {
        "properties": {
          "html": {
            "maxLength": 5242880,
            "title": "Html",
            "type": "string"
          },
          "page_num": {
            "default": 1,
            "maximum": 9999.0,
            "minimum": 1.0,
            "title": "Page Num",
            "type": "integer"
          }
        },
        "required": [
          "html"
        ],
        "title": "HtmlPayload",
        "type": "object"
      },
      "JavdbLoginRefreshPayload": {
        "description": "Optional request body for POST /api/login/refresh.\n\nEmpty body (or omitted body) defaults to proxy_mode='auto', which preserves\nthe original behavior. Existing callers that POST without a body still work.",
        "properties": {
          "max_attempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Attempts"
          },
          "pool_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pool Names"
          },
          "proxy_mode": {
            "default": "auto",
            "enum": [
              "auto",
              "none",
              "single",
              "pool"
            ],
            "title": "Proxy Mode",
            "type": "string"
          },
          "proxy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Url"
          }
        },
        "title": "JavdbLoginRefreshPayload",
        "type": "object"
      },
      "JavdbLoginRefreshResponseV2": {
        "description": "Structured response for POST /api/login/refresh with categorized errors.",
        "properties": {
          "attempts": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Attempts",
            "type": "array"
          },
          "error_category": {
            "anyOf": [
              {
                "enum": [
                  "invalid_credentials",
                  "ip_banned",
                  "captcha_failed",
                  "cloudflare_blocked",
                  "connection_error",
                  "no_credentials",
                  "no_proxy_succeeded",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Category"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "output": {
            "default": "",
            "title": "Output",
            "type": "string"
          },
          "proxy_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Used"
          },
          "status": {
            "enum": [
              "ok",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "JavdbLoginRefreshResponseV2",
        "type": "object"
      },
...
      "LoginPayload": {
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginPayload",
        "type": "object"
      },
      "LoginResponse": {
        "additionalProperties": true,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in",
          "csrf_token",
          "role",
          "username"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "StatusOkResponse": {
        "properties": {
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "StatusOkResponse",
        "type": "object"
      },
...
      "UrlPayload": {
        "properties": {
          "page_num": {
            "default": 1,
            "title": "Page Num",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "use_cf_bypass": {
            "default": true,
            "title": "Use Cf Bypass",
            "type": "boolean"
          },
          "use_cookie": {
            "default": false,
            "title": "Use Cookie",
            "type": "boolean"
          },
          "use_proxy": {
            "default": true,
            "title": "Use Proxy",
            "type": "boolean"
          }
        },
        "required": [
          "url"
        ],
        "title": "UrlPayload",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Fullstack API for config, tasks and parsing.",
    "title": "JAVDB AutoSpider API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/change-password": {
      "post": {
        "operationId": "change_password_api_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password"
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
    "/api/crawl/index": {
      "post": {
        "operationId": "api_crawl_index_api_crawl_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlIndexPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Crawl Index"
      }
    },
    "/api/detect-page-type": {
      "post": {
        "operationId": "api_detect_page_type_api_detect_page_type_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Detect Page Type"
      }
    },
...
    "/api/login/refresh": {
      "post": {
        "operationId": "refresh_javdb_session_api_login_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JavdbLoginRefreshPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JavdbLoginRefreshResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Javdb Session"
      }
    },
...
    "/api/parse/category": {
      "post": {
        "operationId": "api_parse_category_api_parse_category_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Category"
      }
    },
    "/api/parse/detail": {
      "post": {
        "operationId": "api_parse_detail_api_parse_detail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Detail"
      }
    },
    "/api/parse/index": {
      "post": {
        "operationId": "api_parse_index_api_parse_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Index"
      }
    },
    "/api/parse/tags": {
      "post": {
        "operationId": "api_parse_tags_api_parse_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Tags"
      }
    },
    "/api/parse/top": {
      "post": {
        "operationId": "api_parse_top_api_parse_top_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Top"
      }
    },
    "/api/parse/url": {
      "post": {
        "operationId": "api_parse_url_api_parse_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Parse Url"
      }
    },
...
  }
}